feat: add export and importing for custom types, move @types/react-modal to dependencies
