+++ package.json
@@ -73,6 +73,7 @@
         "rc-slider": "9.7.5",
         "react-datepicker": "4.5.0",
         "react-modal": "~3.11.2",
+        "@types/react-modal": "3.10.6",
         "react-popper": "2.2.5",
         "react-virtualized": "~9.22.0",
         "remove-accents": "~0.4.2",
@@ -108,7 +109,6 @@
         "@types/rc-slider": "9.3.1",
         "@types/react": "16.14.34",
         "@types/react-datepicker": "4.3.3",
-        "@types/react-modal": "3.10.6",
         "@types/react-virtualized": "9.21.21",
         "@typescript-eslint/eslint-plugin": "4.33.0",
         "@typescript-eslint/parser": "4.33.0",
