+++ src/components/Callout/Callout.tsx
@@ -3,6 +3,7 @@ import { MdClose } from 'react-icons/md';
 import { bem } from '../../utils';
 import { Context } from '../../constants';
 import styles from './Callout.scss';
+import { NotEmptyReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Content to be rendered inside the container */
