+++ src/components/Dropdown/Dropdown.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { useSelect } from 'downshift';
 import { usePopper } from 'react-popper';
-import { EmptyElement } from '@textkernel/oneui/customTypes/customTypes';
+import { EmptyElement, NotEmptyReactNode } from '../../customTypes/types';
 import { bem } from '../../utils/bem/bem';
 import { mergeRefs } from '../../utils/mergeRefs';
 import { PopupPlacement } from '../../constants';
