+++ src/components/Footer/Footer.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { ReactNode } from '../../customTypes/types';
 import { bem } from '../../utils';
 import { BlockWidthRestrictor } from '../WidthRestrictor';
 import styles from './Footer.scss';
