+++ src/components/Header/Header.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { bem } from '../../utils';
 import { BlockWidthRestrictor } from '../WidthRestrictor';
 import styles from './Header.scss';
+import { ReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** supply a logo ready to be rendered or attributes to build a simple one. */
