+++ src/components/Highlighter/StringHighlighter/StringHighlighter.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { ReactNode } from '../../../customTypes/types';
 import {
     HighlighterCore,
     HighlighterCoreTerm,
