+++ src/components/LoadingSpinner/LoadingSpinner.tsx
@@ -3,6 +3,7 @@ import { bem } from '../../utils';
 import { Text } from '../Text';
 import styles from './LoadingSpinner.scss';
 import { Context } from '../../constants';
+import { ReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Center the spinner relative to parent element or viewport */
