+++ src/components/Navigation/NavBar/NavBar.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { bem } from '../../../utils';
 import { BlockWidthRestrictor } from '../../WidthRestrictor';
 import styles from './NavBar.scss';
+import { ReactNode } from '../../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Node(s) to be rendered as navigation */
