+++ src/components/Pill/PillDropdown/PillDropdown.tsx
@@ -2,6 +2,7 @@ import * as React from 'react';
 import { bem } from '../../../utils';
 import { Button } from '../../Buttons';
 import styles from './PillDropdown.scss';
+import { ReactNode } from '../../../customTypes/types';
 
 export interface PillDropdownChildrenParams {
     /** function that closes the dropdown */
