+++ src/components/StickyHeader/StickyHeader.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { bem } from '../../utils';
 import classnames from './StickyHeader.scss';
+import { SingleReactNode } from '../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Node to be rendered as a sticky header */
