+++ src/components/TwoPaneView/LeftPane/LeftPane.tsx
@@ -1,6 +1,7 @@
 import React, { forwardRef } from 'react';
 import { bem } from '../../../utils';
 import styles from './LeftPane.scss';
+import { ReactNode } from '../../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Node(s) to be rendered in the left pane */
