+++ src/components/WidthRestrictor/PageWidthRestrictor/PageWidthRestrictor.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { bem } from '../../../utils';
 import styles from './PageWidthRestrictor.scss';
+import { NotEmptyReactNode } from '../../../customTypes/types';
 
 export interface Props extends React.HTMLAttributes<HTMLDivElement> {
     /** Node(s) to be rendered inside the container */
