--- src/customTypes/customTypes.ts
@@ -1,1 +1,0 @@
-export type EmptyElement = false | null | undefined;
